# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: MS Scheduling
  description: Microsserviço de agendamentos da barbearia
  version: 1.0.0
servers:
  - url: http://localhost:8082/api/barbershop
paths:
  /scheduling/{id}:
    get:
      tags:
        - Scheduling
      summary: Buscar agendamento por ID
      operationId: findById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulingDTO'
    put:
      tags:
        - Scheduling
      summary: Atualizar agendamento
      operationId: update
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchedulingDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulingDTO'
    delete:
      tags:
        - Scheduling
      summary: Excluir agendamento
      operationId: delete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: string
  /scheduling:
    get:
      tags:
        - Scheduling
      summary: Listar agendamentos
      operationId: findAll
      parameters:
        - name: schedulingFilter
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/SchedulingFilter'
        - name: pageable
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageableDTO'
    post:
      tags:
        - Scheduling
      summary: Cadastrar agendamento
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchedulingDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SchedulingDTO'
components:
  schemas:
    SchedulingDTO:
      required:
        - barberId
        - customerId
        - dateTime
      type: object
      properties:
        customerId:
          type: integer
          format: int64
        barberId:
          type: integer
          format: int64
        status:
          type: string
          enum:
            - AGENDADO
            - AGUARDANDO_PAGAMENTO
        dateTime:
          type: string
          format: date-time
        service:
          type: string
          enum:
            - CABELO
            - BARBA
            - CABELO_BARBA
    SchedulingFilter:
      type: object
      properties:
        customerId:
          type: integer
          format: int64
        barberId:
          type: integer
          format: int64
        dateTime:
          type: string
          format: date-time
        service:
          type: string
          enum:
            - CABELO
            - BARBA
            - CABELO_BARBA
    Pageable:
      type: object
      properties:
        page:
          minimum: 0
          type: integer
          format: int32
        size:
          minimum: 1
          type: integer
          format: int32
        sort:
          type: array
          items:
            type: string
    PageableDTO:
      type: object
      properties:
        numberOfElements:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        schedulingList:
          type: array
          items:
            $ref: '#/components/schemas/Scheduling'
    Scheduling:
      type: object
      properties:
        id:
          type: integer
          format: int64
        customerId:
          type: integer
          format: int64
        barberId:
          type: integer
          format: int64
        status:
          type: string
          enum:
            - AGENDADO
            - AGUARDANDO_PAGAMENTO
        dateTime:
          type: string
          format: date-time
        service:
          type: string
          enum:
            - CABELO
            - BARBA
            - CABELO_BARBA